Add new implementation for Entra externalId
