+++ src/LeaveSystem.Functions/LeaveSystem.Functions.Roles/GetRolesOld.cs
@@ -1,0 +1,94 @@
+using System.Text;
+using System.Text.Json;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.Cosmos;
+using Microsoft.Azure.Cosmos.Linq;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Extensions.Logging;
+
+namespace LeaveSystem.Functions
+{
+    public class GetRolesOld
+    {
+        private static readonly JsonSerializerOptions JsonSerializerOptions = new()
+        {
+            PropertyNameCaseInsensitive = true
+        };
+        private static readonly CosmosLinqSerializerOptions LinqSerializerOptions = new()
+        {
+            PropertyNamingPolicy = CosmosPropertyNamingPolicy.CamelCase
+        };
+        private readonly ILogger<GetRolesOld> logger;
+
+        public GetRolesOld(ILogger<GetRolesOld> logger) => this.logger = logger;
+
+        [Function(nameof(GetRolesOld))]
+        public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", "post")] HttpRequest req,
+            [CosmosDBInput(
+                databaseName: "LeaveSystem",
+                containerName: "Roles",
+                Connection = "CosmosDBConnection")] Container container)
+        {
+            logger.LogInformation("C# HTTP trigger function processed a request.");
+            if (!ValidateToken(req.Headers.Authorization))
+            {
+                logger.LogWarning("Wrong RestApi credentials.");
+                return new ObjectResult(new
+                {
+                    version = "1.0",
+                    status = 401,
+                    code = "errorCode",
+                    requestId = "requestId",
+                    userMessage = "Wrong RestApi credentials. Contact with administrators.",
+                    developerMessage = $"The provided code auth {req.Headers.Authorization} does not match the expected login and password."
+                })
+                { StatusCode = 401 };
+            }
+            var userIdModel = await JsonSerializer.DeserializeAsync<UserId>(req.Body, JsonSerializerOptions);
+            if (userIdModel is null)
+            {
+                return new NotFoundObjectResult("userId");
+            }
+
+            var userId = userIdModel.Id;
+            this.logger.LogInformation("Deserialized user id {UserId}.", userId);
+            var iterator = container.GetItemLinqQueryable<RolesModel>(linqSerializerOptions: LinqSerializerOptions)
+                .Where(r => r.Id == userId)
+                .ToFeedIterator();
+            var roles = iterator.HasMoreResults ?
+                (await iterator.ReadNextAsync()).FirstOrDefault()?.Roles :
+                null;
+            return new OkObjectResult(new { roles = roles ?? Enumerable.Empty<string>() });
+        }
+        private bool ValidateToken(string? header)
+        {
+            //Checking the header
+            if (!string.IsNullOrEmpty(header) && header.StartsWith("Basic", StringComparison.Ordinal))
+            {
+                //Extracting credentials
+                // Removing "Basic " Substring
+                var encodedUsernamePassword = header["Basic ".Length..].Trim();
+                //Decoding Base64
+                var encoding = Encoding.GetEncoding("iso-8859-1");
+                var usernamePassword = encoding.GetString(Convert.FromBase64String(encodedUsernamePassword));
+                //Splitting Username:Password
+                var seperatorIndex = usernamePassword.IndexOf(':');
+                // Extracting the individual username and password
+                var username = usernamePassword[..seperatorIndex];
+                var password = usernamePassword[(seperatorIndex + 1)..];
+                //Validating the credentials
+                return username == Environment.GetEnvironmentVariable("RestApiUsername") &&
+                    password == Environment.GetEnvironmentVariable("RestApiPassword");
+            }
+            else
+            {
+                logger.LogWarning("Header is missing");
+                return false;
+            }
+        }
+    }
+
+    public record UserId(Guid Id);
+    public record RolesModel(Guid Id, string[] Roles);
+}
